WIP Update documentation for VMware Fusion Setup
Without this patch it's difficult to get a Razor Server setup and
running in a development environment without knowing the details of TFTP
and VMware Fusion DHCP configuration.

This patch addresses the problem by adding some helpful information on
configuration file locations and TFTP server configurations for use on a
Mac OS X development workstation.